requirements.txt

easyocr==1.7.2
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84

PyYAML==6.0.2
scikit_learn==1.2.2
torch==2.4.1
ultralytics==8.2.103
